# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HonMoon API",
    "description": "HonMoon API 문서",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/missions/{id}/submit-quiz": {
      "post": {
        "tags": [
          "Mission Detail"
        ],
        "summary": "퀴즈 답변 제출 (통합)",
        "description": "현재 로그인한 사용자가 미션 퀴즈 답변(텍스트/객관식/이미지)을 제출합니다.",
        "operationId": "submitQuizAnswer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "미션 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "textAnswer",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "selectedChoiceIndex",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "uploadedImageUrl",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyUserActivityResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/missions/{id}/submit-quiz/me": {
      "post": {
        "tags": [
          "Mission Detail"
        ],
        "summary": "내 퀴즈 답변 제출 (통합)",
        "description": "현재 로그인한 사용자가 미션 퀴즈 답변을 제출합니다.",
        "operationId": "submitQuizAnswerMe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "미션 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "textAnswer",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "selectedChoiceIndex",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "uploadedImageUrl",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyUserActivityResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/missions/{id}/submit-image-answer": {
      "post": {
        "tags": [
          "Mission Detail"
        ],
        "summary": "미션 이미지 답변 제출",
        "description": "이미지 업로드 퀴즈 미션에 이미지를 제출합니다. 서버가 AI로 텍스트를 추출해 정답을 판별합니다.",
        "operationId": "submitMissionImageAnswer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "미션 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionImageAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyMissionAnswerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/missions/{id}/submit-answer": {
      "post": {
        "tags": [
          "Mission Detail"
        ],
        "summary": "미션 답변 제출",
        "description": "퀴즈 미션에 텍스트 답변을 제출합니다. 정답이면 포인트가 지급됩니다.",
        "operationId": "submitMissionAnswer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "미션 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyMissionAnswerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/missions/{id}/image/upload-url": {
      "post": {
        "tags": [
          "Mission Detail"
        ],
        "summary": "미션 이미지 업로드 URL 발급",
        "description": "미션 관련 이미지를 버킷에 업로드하기 위한 Presigned URL을 발급합니다.",
        "operationId": "createMissionImageUploadUrl",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "fileName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contentType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyPresignedUrlResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/missions/{id}": {
      "get": {
        "tags": [
          "Mission Detail"
        ],
        "summary": "미션 상세 정보 조회",
        "description": "특정 미션의 상세 정보를 조회합니다.",
        "operationId": "getMissionDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "미션 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MissionDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/missions/{id}/detail": {
      "get": {
        "tags": [
          "Mission Detail"
        ],
        "summary": "미션 상세 정보 조회(별칭)",
        "description": "특정 미션의 상세 정보를 조회합니다. 기존 경로와 동일한 응답을 반환합니다.",
        "operationId": "getMissionDetailAlias",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "미션 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MissionDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserPrincipal": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "provider",
          "roles",
          "subject"
        ]
      },
...
      "AnswerCheckResult": {
        "type": "object",
        "properties": {
          "isCorrect": {
            "type": "boolean"
          },
          "confidence": {
            "type": "number",
            "format": "double"
          },
          "reasoning": {
            "type": "string"
          },
          "extractedText": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          }
        },
        "required": [
          "confidence",
          "isCorrect",
          "reasoning"
        ]
      },
      "UserActivityResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "placeId": {
            "type": "integer",
            "format": "int64"
          },
          "missionId": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "isCompleted": {
            "type": "boolean"
          },
          "pointsEarned": {
            "type": "integer",
            "format": "int32"
          },
          "textAnswer": {
            "type": "string"
          },
          "selectedChoiceIndex": {
            "type": "integer",
            "format": "int32"
          },
          "uploadedImageUrl": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          },
          "aiResult": {
            "$ref": "#/components/schemas/AnswerCheckResult"
          },
          "alreadyExists": {
            "type": "boolean"
          }
        },
        "required": [
          "alreadyExists",
          "createdAt",
          "id",
          "isCompleted",
          "modifiedAt",
          "placeId",
          "pointsEarned",
          "userId"
        ]
      },
...
      "ResponseBodyUserActivityResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserActivityResponse"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "status"
        ]
      },
      "MissionImageAnswerRequest": {
        "type": "object",
        "properties": {
          "imageUrl": {
            "type": "string"
          }
        },
        "required": [
          "imageUrl"
        ]
      },
      "MissionAnswerResponse": {
        "type": "object",
        "properties": {
          "isCorrect": {
            "type": "boolean"
          },
          "pointsEarned": {
            "type": "integer",
            "format": "int32"
          },
          "explanation": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          },
          "aiResult": {
            "$ref": "#/components/schemas/AnswerCheckResult"
          }
        },
        "required": [
          "isCorrect",
          "pointsEarned"
        ]
      },
      "ResponseBodyMissionAnswerResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MissionAnswerResponse"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "status"
        ]
      },
      "MissionAnswerRequest": {
        "type": "object",
        "properties": {
          "answer": {
            "type": "string"
          }
        },
        "required": [
          "answer"
        ]
      },
      "PresignedUrlResponse": {
        "type": "object",
        "properties": {
          "uploadUrl": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "expiresAt",
          "fileName",
          "uploadUrl"
        ]
      },
      "ResponseBodyPresignedUrlResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PresignedUrlResponse"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "status"
        ]
      },
...
      "MissionChoicesVo": {
        "type": "object",
        "properties": {
          "choices": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "choices"
        ]
      },
      "MissionDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "points": {
            "type": "integer",
            "format": "int32"
          },
          "missionType": {
            "type": "string",
            "enum": [
              "QUIZ_MULTIPLE_CHOICE",
              "QUIZ_TEXT_INPUT",
              "QUIZ_IMAGE_UPLOAD",
              "PLACE_VISIT",
              "PHOTO_UPLOAD",
              "SURVEY"
            ]
          },
          "placeId": {
            "type": "integer",
            "format": "int64"
          },
          "question": {
            "type": "string"
          },
          "answer": {
            "type": "string"
          },
          "choices": {
            "$ref": "#/components/schemas/MissionChoicesVo"
          },
          "answerExplanation": {
            "type": "string"
          },
          "correctImageUrl": {
            "type": "string"
          },
          "imageUploadInstruction": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "createdAt",
          "id",
          "missionType",
          "modifiedAt",
          "points",
          "title"
        ]
      },
...
    }
  }
}